feat: get config from db before import & fix some  issues
