--- app/pages/Import/TaskCreate/FileSelect/index.module.less
@@ -1,11 +1,0 @@
-.popoverFileSelect {
-	margin-bottom: 20px;
-	:global(.ant-popover-title) {
-		border-bottom: none;
-	}
-}
-.fileSelectForm {
-	.fileSelect {
-		width: 280px;
-	}
-}
