+++ server/api/studio/restapi/import.api
@@ -76,7 +76,7 @@ type (
 	ImportTaskConfig {
 		Client  Client    `json:"client" validate:"required"`
 		Manager Manager   `json:"manager" validate:"required"`
-		Sources []Sources `json:"sources" validate:"required"`
+		Sources []*Source `json:"sources" validate:"required"`
 		Log     *Log      `json:"log,omitempty,optional"`
 	}
 
@@ -99,13 +99,15 @@ type (
 		StatsInterval       *string `json:"statsInterval,omitempty,optional"`
 	}
 
-	Sources {
-		CSV   ImportTaskCSV `json:"csv" validate:"required"`
-		Path  string        `json:"path,optional,omitempty"`
-		S3    *S3Config     `json:"s3,optional,omitempty"`
-		SFTP  *SFTPConfig   `json:"sftpConfig,optional,omitempty"`
-		Tags  []Tag         `json:"tags,optional"`
-		Edges []Edge        `json:"edges,optional"`
+	Source {
+		CSV                ImportTaskCSV `json:"csv" validate:"required"`
+		Path               string        `json:"path,optional,omitempty"`
+		S3                 *S3Config     `json:"s3,optional,omitempty"`
+		SFTP               *SFTPConfig   `json:"sftp,optional,omitempty"`
+		DatasourceId       *int64        `json:"datasourceId,optional,omitempty"`
+		DatasourceFilePath *string       `json:"datasourceFilePath,optional,omitempty"`
+		Tags               []Tag         `json:"tags,optional"`
+		Edges              []Edge        `json:"edges,optional"`
 	}
 
 	Log {
